\name{rmongoshell-package}
\alias{rmongoshell-package}
\alias{rmongoshell}
\docType{package}
\title{
mongo shell R connection
}
\description{
A lightweight node.js based tool to query mongodb databases with minimal overhead. Less memory intensive than the more complex alternatives (I'm looking at you Java).
}
\details{
\tabular{ll}{
Package: \tab rmongoshell\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-13\cr
License: \tab Apache License, Version 2.0\cr
}
Make sure node is in your path, blah blah blah
}
\author{
Bill Beesley <bill@billbeesley.com>
}
\references{
\url{http://mongodb.github.io/node-mongodb-native/}
}
\keyword{ package }
\keyword{ database }
\examples{
\dontrun{
plot(x)
}
}
